% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-CtoTAssignment.R
\name{CtoTAssignment}
\alias{CtoTAssignment}
\alias{CtoTAssignment,GRanges-method}
\title{CtoTAssignment for assigning the truncation read counts}
\usage{
CtoTAssignment(object, bw_positive, bw_negative, sampleName = "")

\S4method{CtoTAssignment}{GRanges}(
    object, bw_positive, bw_negative, sampleName = "")
}
\arguments{
\item{object}{A GRanges object which should contains all the single
nucleotide peaks of miCLIP2 experiment.}

\item{bw_positive}{A path to the big wig file of C to T transition read
counts at the positive strand that output from the preprocess in the
m6Aboost pipeline.}

\item{bw_negative}{A path to the big wig file of C to T transition read
counts at the negative strand that output from the preprocess in the
m6Aboost pipeline.}

\item{sampleName}{The sample name of the big wig file.}
}
\value{
A GRanges object with the truncation read counts.
}
\description{
An function for assigning the CtoT transition read counts from
    bigwig.
}
\examples{
    testpath <- system.file("extdata", package = "m6Aboost")
    test <- readRDS(file.path(testpath, "test.rds"))
    positiveBW <- file.path(testpath, "test_positive.bw")
    negativeBW <- file.path(testpath, "test_negative.bw")

    test <- truncationAssignment(test, positiveBW, negativeBW, "CtoT1")
}
\author{
You Zhou
}
