% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures,GRanges-method}
\title{getFeatures for the miCILP2 data}
\usage{
getFeatures(object, annotation, colname_reads = "")

\S4method{getFeatures}{GRanges}(object, annotation, colname_reads = "")
}
\arguments{
\item{object}{A GRanges object which should contains all the single
nucleotide peaks of miCLIP2 experiment.}

\item{annotation}{A path way to the annotation file. The format of the
annotation file should be gff3 and downloaded from
https://www.gencodegenes.org/}

\item{colname_reads}{The name of the meta data column which contains the
mean value of the truncation reads number without C to T transition
reads.}
}
\value{
A GRanges object with all the information that required by the
    m6Aboost model.
}
\description{
An function for calculating the relative signal strength and
    assign the features for the m6Aboost model
}
\author{
You Zhou, Kathi Zarnack
}
