% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-m6Aboost.R
\name{m6Aboost}
\alias{m6Aboost}
\alias{m6Aboost,GRanges-method}
\title{getFeatures for the miCILP2 data}
\usage{
m6Aboost(object, genome = "", normalization = TRUE)

\S4method{m6Aboost}{GRanges}(object, genome = "", normalization = TRUE)
}
\arguments{
\item{object}{A GRanges object which should contains all the single
nucleotide peaks of miCLIP2 experiment.}

\item{genome}{The name of the BSgenome that you are working with. For
example "BSgenome.Mmusculus.UCSC.mm10".}

\item{normalization}{A logical vector which indicates whether you would like
normalize the RSS and C to T reads number to the mean value of our
training set. This will help to reduce the false positive rate.}
}
\value{
A GRanges object with all the information that required by the
    m6Aboost model.
}
\description{
An function for calculating the relative signal strength and
    assign the features for the m6Aboost model
}
\author{
You Zhou, Kathi Zarnack
}
